Migrate 404 error to not found page
